TriggerIngestion/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
dnspython==2.3.0
email-validator==2.0.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
multidict==6.0.4
notion-sdk==0.7.0
openai==0.27.4
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
sniffio==1.3.0
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
